branch

+++ pyewts/pyewts.py
@@ -1158,7 +1158,10 @@ class pyewts(object):
                 cc = cc.replace('\u2018', '\'')
                 cc = cc.replace('\u2019', '\'')
                 expect_key = self.ambiguous_key(cc)
-                if expect_key != None and expect_key.intValue() != root_idx:
+                # bug fix by riggy2013, 11/06/2019
+                # invValue is not a valid function in python
+                # if expect_key != None and expect_key.intValue() != root_idx:
+                if expect_key != None and expect_key != root_idx:
                     warns.append("Syllable should probably be \"" + self.ambiguous_wylie(cc) + "\".")
         ret = self.WylieTsekbar()
         ret.uni_string = out
